+++ core/src/main/scala/com/stratio/crossdata/util/utils.scala
@@ -1,0 +1,34 @@
+/*
+ * Copyright (C) 2015 Stratio (http://stratio.com)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.stratio.crossdata.util
+
+import scala.util.Try
+
+
+object using {
+
+  type AutoClosable = {def close(): Unit}
+
+  def apply[A <: AutoClosable, B](resource: A)(code: A => B): B =
+    try {
+      code(resource)
+    }
+    finally {
+      Try(resource.close())
+    }
+
+}
+
